Fixes required for Jupyter Notebooks conversion to PDF
Jupyter comes in with a nice CLI that may perform conversions of a
Notebook to various formats, including HTML and PDF. This conver-
-sion should ideally be automated at build time so that the work
assignments would be readily available in a web browser with nothing
to install (since most undergrads struggle with VSCode or Binder).

However, Jupyter uses a stricter parser. Thus, some code accepted
by VSCode and Binder may lead to issues when using Jupyter directly.

In particular, Markdown bullet points and LaTeX maths formulas are
parsed with a different engine that does not interpret them in the
same fashion.

```bash
jupyter nbconvert --to pdf homework-03.ipynb

jupyter nbconvert --to latex homework-03.ipynb
```

Then add the following snippet at the beginning of the article, just
after the 1st line:

```latex
\usepackage[utf8]{inputenc}
\usepackage[french,english]{babel}
```

You may also change the `\title` and add `\date{}` to remove the date
that appears by default as a subtitle.

Finally, use LaTeX (or XeLaTeX) to generate the PDF.