Update the dependencies to their 2024 latest released versions
I observed on the production site that the emoticons were no longer
displayed. I was unaware of the regression and I do not really know
what the root cause could be. Maybe some update in the unicode
specification and its implementation in web browsers, or maybe some
update to the emoticons themselves.

On the local development system, the update of the dependencies
solved the display of the emoticons. Hopefully, it shall fix it in
production as well.

I tried to update the dependencies last year and faced an error
message at compile time. Apparently, it was fixed in the meantime,
which allows me to push this simple update today.

It seems that even statically generated websites need updates to
deal with regressions.